+++ R/derive_date_vars.R
@@ -901,8 +901,9 @@ derive_vars_dtm <- function(dataset,
     date_imputation = date_imputation,
     time_imputation = time_imputation,
     min_dates = lapply(min_dates, eval_tidy, data = mask),
-    max_dates = lapply(min_dates, eval_tidy, data = mask),
+    max_dates = lapply(max_dates, eval_tidy, data = mask),
     preserve = preserve
+
   )
 
   if (flag_imputation %in% c("both", "date") ||