+++ R/iso_dtm.R
@@ -3,6 +3,9 @@ as_iso_dtm <- function(x, time_zone = "UTC") {
     dtm <- ymd_hms(x, tz = time_zone)
   } else {
     dtm <- x
+    if (is.null(attr(dtm, "tzone")) || attr(dtm, "tzone") == "") {
+      attr(dtm, "tzone") <- time_zone
+    }
   }
   class(dtm) <- union("iso_dtm", class(dtm))
   dtm