+++ NEWS.md
@@ -15,13 +15,21 @@
 dataset](https://pharmaverse.github.io/admiral/articles/queries_dataset.html) required as input for
 `derive_vars_query()` (#606)
 
+- `create_single_dose_dataset()` - Derives dataset of single dose from aggregate dose information (#660)
+
 - New functions for deriving first or last dates from multiple source datasets (#753):
 
   - `derive_var_extreme_dtm()` - Derive First or Last Datetime from Multiple Sources
   - `derive_var_extreme_dt()` - Derive First or Last Date from Multiple Sources
 
+- New function `derive_extreme_records()` for adding the first or last
+observation within each by group to the dataset (#1042)
+
 - `derive_var_shift()` - Derives a character shift variable containing concatenated shift in 
 values based on user-defined pairing (#944)
+
+- `derive_var_analysis_ratio` - Derives a ratio variable based on user-supplied variables
+from a BDS dataset, e.g. ADLB. (#943)
 
 - New high order functions (#701):
 
@@ -71,6 +79,7 @@ Address [CRAN comments](https://github.com/pharmaverse/admiral/issues/918) raise
 - `derive_var_ontrtfl()` now has a `new_var` parameter to support the derivation of `ONTRxxFL` and `ONTRTwFL` variables (#721)
 
 - `derive_vars_dtm()`, `derive_var_disposition` and `derive_var_lstalvdt` now have `preserve` argument.  A user can preserve partial dates when doing date imputation, e.g. `2019---07` would become `2019-06-07` by setting `preserve` to `TRUE` when doing date_imputation (#592)
+
 
 - `derive_vars_dtm()` now has `ignore_seconds_flag` argument so users can suppress `"S"` flag if seconds are not recorded in the data (#589)
 
@@ -183,6 +192,8 @@ to specify the unit of the input age (#569)
 ## Bug Fixes
 
 - `derive_vars_dtm()` no longer shifts the time of the input `--DTC` variable (#436)
+
+- `derive_vars_dtm()` Change the min_dates with max_dates in the lapply statement when computing max_dates (#687)
 
 ## Documentation